Issue #20: removed total input length limitation
Rationale: the total input length limitation in SHA-256 is (2^64 - 1)
bits. While the streaming API would in principle let client code reach
that limit, since it may invoke `sha_256_write` as many times as it
wants, in practice, before this commit, the limit was `MAX_SIZE` bytes,
which could be much less than the SHA-256 limit. I.e. before this
commit, beyond that total input length, the behavior was undefined.

This commit removes that limitation, at the cost of requiring compiler
support for `uint64_t`, which should not be a problem, since this
implementation already required `uint32_t` support. Also, theoretically,
there could be a performance impact, but since this change only impacts
total length calculation and encoding, this should not be noticeable.

For the sake of completion, it should be noted that the limitation
removed by this commit only impacted the streaming API, not the simpler
`calc_sha_256` API, since for the latter, client code had to provide a
single input data buffer, whose length, by definition of `size_t`, could
not be greater than `MAX_SIZE`.

+++ sha-256.c
@@ -176,8 +176,8 @@ uint8_t *sha_256_close(struct Sha_256 *sha_256)
 	uint32_t *const h = sha_256->h;
 
 	/*
-	 * The current chunk cannot be full. Otherwise, it would already have been consumed. I.e. there is space left for
-	 * at least one byte. The next step in the calculation is to add a single one-bit to the data.
+	 * The current chunk cannot be full. Otherwise, it would already have been consumed. I.e. there is space left
+	 * for at least one byte. The next step in the calculation is to add a single one-bit to the data.
 	 */
 	*pos++ = 0x80;
 	--space_left;
@@ -196,7 +196,7 @@ uint8_t *sha_256_close(struct Sha_256 *sha_256)
 	const size_t left = space_left - TOTAL_LEN_LEN;
 	memset(pos, 0x00, left);
 	pos += left;
-	size_t len = sha_256->total_len;
+	uint64_t len = sha_256->total_len;
 	pos[7] = (uint8_t)(len << 3);
 	len >>= 5;
 	int i;
